# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  version: 1.0.0
  description: API for WASAText instant messaging application.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /session:
    post:
      tags:
        - login
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created, and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Maria
                  pattern: "^.*?$"
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    type: string
                    example: "abcdef012345"

  /users/me/name:
    put:
      tags:
        - users
      summary: Set my username
      operationId: setMyUserName
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '204':
          description: Username updated successfully
        '400':
          description: Invalid username
        '409':
          description: Username already in use

  /users/me/photo:
    put:
      tags:
        - users
      summary: Set my profile photo
      operationId: setMyPhoto
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        '204':
          description: Photo updated successfully

  /users:
    get:
      tags:
        - users
      summary: Search for users
      operationId: searchUsers
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Username query string
          schema:
            type: string
      responses:
        '200':
          description: List of users matching the query
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

  /conversations:
    get:
      tags:
        - conversations
      summary: Get my conversations
      description: Returns a list of conversations sorted in reverse chronological order.
      operationId: getMyConversations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                items:
                  $ref: '#/components/schemas/ConversationUnread'

    post:
      tags:
        - conversations
      summary: Create a new conversation (group)
      operationId: createConversation
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Group name (required for groups)
                members:
                  type: array
                  items:
                    type: string
                  description: Initial members (optional)
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetails'

  /conversations/{conversationId}:
    get:
      tags:
        - conversations
      summary: Get a conversation
      description: Get conversation details and messages.
      operationId: getConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation details and messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetails'
        '404':
          description: Conversation not found

  /conversations/{conversationId}/name:
    put:
      tags:
        - conversations
      summary: Set group name
      operationId: setGroupName
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
              required:
                - name
      responses:
        '204':
          description: Group name updated
        '403':
          description: Not allowed (not a group or not a member)

  /conversations/{conversationId}/photo:
    put:
      tags:
        - conversations
      summary: Set group photo
      operationId: setGroupPhoto
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        '204':
          description: Group photo updated
        '403':
          description: Not allowed

  /conversations/{conversationId}/members/{userId}:
    put:
      tags:
        - conversations
      summary: Add a user to a group
      operationId: addToGroup
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User added to group
        '403':
          description: Not allowed
        '404':
          description: User or conversation not found

  /conversations/{conversationId}/members/me:
    delete:
      tags:
        - conversations
      summary: Leave a group
      operationId: leaveGroup
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Left group successfully

  /conversations/{conversationId}/messages:
    post:
      tags:
        - messages
      summary: Send or forward a message
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/NewTextMessage'
                - $ref: '#/components/schemas/NewPhotoMessage'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /conversations/{conversationId}/messages/forwarded:
    post:
      tags:
          - messages
      summary: Forward a message
      operationId: forwardMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                forwardedMessageId:
                  type: string
              required:
                - forwardedMessageId
      responses:
        '201':
          description: Message forwarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /conversations/{conversationId}/messages/{messageId}:
    delete:
      tags:
        - messages
      summary: Delete a message
      operationId: deleteMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Message deleted

  /conversations/{conversationId}/messages/{messageId}/reactions:
    post:
      tags:
        - reactions
      summary: React to a message (comment)
      operationId: commentMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emoji:
                  type: string
              required:
                - emoji
      responses:
        '201':
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'

  /conversations/{conversationId}/messages/{messageId}/reactions/{reactionId}:
    delete:
      tags:
        - reactions
      summary: Remove a reaction (uncomment)
      operationId: uncommentMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
        - name: reactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Reaction removed

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        photo:
          type: string
          format: byte
          nullable: true

    ConversationUnread:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        isGroup:
          type: boolean
        photo:
          type: string
          format: byte
          nullable: true
        lastMessage:
          $ref: '#/components/schemas/Message'
        unreadCount:
          type: integer

    ConversationDetails:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        isGroup:
          type: boolean
        ownerId:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'

    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [text, image]
        timestamp:
          type: string
          format: date-time
        received:
          type: boolean
        read:
          type: boolean
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/Reaction'

    NewTextMessage:
      type: object
      properties:
        type:
          type: string
          enum: [text]
        text:
          type: string

    NewPhotoMessage:
      type: object
      properties:
        type:
          type: string
          enum: [image]
        image:
          type: string
          format: binary

    Reaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        emoji:
          type: string